Fix: entity 수정
